Complete rework of the key data structure + popups
This rework is a very important one, as it solves many conflicts which
arised while implementing special keyboard layouts like dvorak, etc.

With the new data structure, popup keys carry around way less baggage,
but at the same time the structure around the keys have improved a lot and provide more useful information for the popup manager.
This includes that smart hint/accent prioritization is now working
theoretically, now it just needs to be defined correctly in the
json files.

Also the layout and extended popup json files got a authors field
which indicates who the author is. At the current time not very
important, but this is a small step in modularization into assets,
which are the very base of extension support (but that's still a
looong way).

Also some parts of the LayoutManager got a code cleanup, which makes
especially the merging of popups and hints better readable.